- Added Page Routing - Added template for pages - Added Top Navbar
