+++ package.json
@@ -25,6 +25,7 @@
     "postcss": "^8.4.24",
     "prettier": "^2.8.8",
     "prettier-plugin-tailwindcss": "^0.3.0",
+    "react-router-dom": "^6.12.1",
     "tailwindcss": "^3.3.2",
     "vite": "^4.3.9"
   }
