+++ src/index.css
@@ -5,6 +5,7 @@
 #root{
     height: 100vh;
     width: 100vw;
+    display: flex;
 }
 
 body {
