+++ src/pages/Calendar.jsx
@@ -1,0 +1,12 @@
+import '../App.css'
+
+function Calendar() {
+
+  return (
+    <div className='w-full h-full bg-zinc-300'>
+      <h1 className='flex h-24 items-center justify-center text-3xl'>Calendar</h1>
+    </div>
+  )
+}
+
+export default Calendar
