+++ src/pages/Planner.jsx
@@ -1,0 +1,12 @@
+import '../App.css'
+
+function Planner() {
+
+  return (
+    <div className='w-full h-full bg-slate-300'>
+      <h1 className='flex h-24 items-center justify-center text-3xl'>Planner</h1>
+    </div>
+  )
+}
+
+export default Planner
